Massive backend rewrite to use Thrift RPC instead of gRPC.

+++ Areas/DHub/Models/IDroneTelemetry.cs
@@ -1,6 +1,6 @@
 namespace DroHub.Areas.DHub.Models {
     public interface IDroneTelemetry {
         string Serial { get; set; }
-        uint Timestamp { get; set; }
+        long Timestamp { get; set; }
     }
 }
